+++ .github/workflows/ci.yml
@@ -63,17 +63,26 @@ jobs:
       fail-fast: false
       matrix:
         info:
+        - name: PS_5.1_x64_Windows
+          psversion: '5.1'
+          os: windows-latest
         - name: PS_7.2_x64_Windows
-          psversion: '7.2.8'
+          psversion: '7.2.0'
           os: windows-latest
         - name: PS_7.2_x64_Linux
-          psversion: '7.2.8'
+          psversion: '7.2.0'
           os: ubuntu-latest
         - name: PS_7.3_x64_Windows
-          psversion: '7.3.1'
+          psversion: '7.3.0'
           os: windows-latest
         - name: PS_7.3_x64_Linux
-          psversion: '7.3.1'
+          psversion: '7.3.0'
+          os: ubuntu-latest
+        - name: PS_7.4_x64_Windows
+          psversion: '7.4.0'
+          os: windows-latest
+        - name: PS_7.4_x64_Linux
+          psversion: '7.4.0'
           os: ubuntu-latest
 
     steps:
