+++ module/PSToml.psd1
@@ -11,10 +11,10 @@
 @{
 
     # Script module or binary module file associated with this manifest.
-    RootModule = 'bin/net6.0/PSToml.dll'
+    RootModule = 'PSToml.psm1'
 
     # Version number of this module.
-    ModuleVersion = '0.2.0'
+    ModuleVersion = '0.3.0'
 
     # Supported PSEditions
     # CompatiblePSEditions = @()
@@ -35,10 +35,10 @@
     Description = 'TOML parser and writer for PowerShell'
 
     # Minimum version of the PowerShell engine required by this module
-    PowerShellVersion = '7.2'
+    PowerShellVersion = '5.1'
 
     # Minimum version of Microsoft .NET Framework required by this module. This prerequisite is valid for the PowerShell Desktop edition only.
-    # DotNetFrameworkVersion = '4.7.2'
+    DotNetFrameworkVersion = '4.7.2'
 
     # Minimum version of the common language runtime (CLR) required by this module. This prerequisite is valid for the PowerShell Desktop edition only.
     # ClrVersion = '4.0'
