--- src/Shared/PSToml.Shared.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Management.Automation" Version="7.2.0" PrivateAssets="all" />
-    <PackageReference Include="Tomlyn" Version="0.16.2" />
-  </ItemGroup>
-</Project>
